Persistence plugin refactor (#26)
No more persistence keys, we're taking advantage of having a type system now.

The persistence plugin has many persistence documents, each of which know how to serialize and unserialize their data. Currently, there are ones for the adventurer's stats, the last saved location (area and marker), and the player's progression through the conversations of the game.

There's definitely still some work to be done on the progression's structure, already the conversation document is getting way too big because its defining the conversations in line with the logic. That definition should be separated out somewhere.